Move ambientindex to environment
