+++ pilot/pkg/bootstrap/server.go
@@ -51,6 +51,7 @@ import (
 	sec_model "istio.io/istio/pilot/pkg/security/model"
 	"istio.io/istio/pilot/pkg/server"
 	"istio.io/istio/pilot/pkg/serviceregistry/aggregate"
+	"istio.io/istio/pilot/pkg/serviceregistry/kube/controller/ambient"
 	"istio.io/istio/pilot/pkg/serviceregistry/provider"
 	"istio.io/istio/pilot/pkg/serviceregistry/serviceentry"
 	"istio.io/istio/pilot/pkg/status"
@@ -295,6 +296,7 @@ func NewServer(args *PilotArgs, initFuncs ...func(*Server)) (*Server, error) {
 	s.initMeshNetworks(args, s.fileWatcher)
 	s.initMeshHandlers(configGen.MeshConfigChanged)
 	s.environment.Init()
+
 	if err := s.environment.InitNetworksManager(s.XDSServer); err != nil {
 		return nil, err
 	}
@@ -324,6 +326,24 @@ func NewServer(args *PilotArgs, initFuncs ...func(*Server)) (*Server, error) {
 
 	if err := s.initControllers(args); err != nil {
 		return nil, err
+	}
+
+	if features.EnableAmbient {
+		e.AmbientIndexes = ambient.New(ambient.Options{
+			Client:          s.kubeClient,
+			SystemNamespace: args.RegistryOptions.KubeOptions.SystemNamespace,
+			DomainSuffix:    e.DomainSuffix,
+			ClusterID:       s.clusterID,
+			Revision:        args.Revision,
+			XDSUpdater:      s.XDSServer,
+			MeshConfig:      e.Watcher,
+			StatusNotifier:  args.RegistryOptions.KubeOptions.StatusWritingEnabled,
+			Debugger:        args.RegistryOptions.KubeOptions.KrtDebugger,
+			Flags: ambient.FeatureFlags{
+				DefaultAllowFromWaypoint:              features.DefaultAllowFromWaypoint,
+				EnableK8SServiceSelectWorkloadEntries: features.EnableK8SServiceSelectWorkloadEntries,
+			},
+		})
 	}
 
 	InitGenerators(s.XDSServer, configGen, args.Namespace, s.clusterID, s.internalDebugMux)
