+++ pilot/pkg/model/context.go
@@ -17,7 +17,6 @@ package model
 import (
 	"encoding/json"
 	"fmt"
-	"net"
 	"regexp"
 	"sort"
 	"strconv"
@@ -30,23 +29,18 @@ import (
 	anypb "google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/structpb"
 
-	meshconfig "istio.io/api/mesh/v1alpha1"
-	"istio.io/istio/pilot/pkg/credentials"
 	"istio.io/istio/pilot/pkg/features"
 	istionetworking "istio.io/istio/pilot/pkg/networking"
-	"istio.io/istio/pilot/pkg/trustbundle"
 	networkutil "istio.io/istio/pilot/pkg/util/network"
 	v3 "istio.io/istio/pilot/pkg/xds/v3"
 	"istio.io/istio/pkg/cluster"
 	"istio.io/istio/pkg/config/constants"
-	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/mesh"
 	"istio.io/istio/pkg/config/mesh/meshwatcher"
 	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/config/schema/kind"
 	"istio.io/istio/pkg/maps"
 	pm "istio.io/istio/pkg/model"
-	"istio.io/istio/pkg/monitoring"
 	"istio.io/istio/pkg/network"
 	"istio.io/istio/pkg/spiffe"
 	"istio.io/istio/pkg/util/identifier"
@@ -97,153 +91,6 @@ func NewEnvironment() *Environment {
 
 // Watcher is a type alias to keep the embedded type name stable.
 type Watcher = meshwatcher.WatcherCollection
-
-// Environment provides an aggregate environmental API for Pilot
-type Environment struct {
-	// Discovery interface for listing services and instances.
-	ServiceDiscovery
-
-	// Config interface for listing routing rules
-	ConfigStore
-
-	// Watcher is the watcher for the mesh config (to be merged into the config store)
-	Watcher
-
-	// NetworksWatcher (loaded from a config map) provides information about the
-	// set of networks inside a mesh and how to route to endpoints in each
-	// network. Each network provides information about the endpoints in a
-	// routable L3 network. A single routable L3 network can have one or more
-	// service registries.
-	NetworksWatcher mesh.NetworksWatcher
-
-	NetworkManager *NetworkManager
-
-	// mutex used for protecting Environment.pushContext
-	mutex sync.RWMutex
-	// pushContext holds information during push generation. It is reset on config change, at the beginning
-	// of the pushAll. It will hold all errors and stats and possibly caches needed during the entire cache computation.
-	// DO NOT USE EXCEPT FOR TESTS AND HANDLING OF NEW CONNECTIONS.
-	// ALL USE DURING A PUSH SHOULD USE THE ONE CREATED AT THE
-	// START OF THE PUSH, THE GLOBAL ONE MAY CHANGE AND REFLECT A DIFFERENT
-	// CONFIG AND PUSH
-	pushContext *PushContext
-
-	// DomainSuffix provides a default domain for the Istio server.
-	DomainSuffix string
-
-	// TrustBundle: List of Mesh TrustAnchors
-	TrustBundle *trustbundle.TrustBundle
-
-	clusterLocalServices ClusterLocalProvider
-
-	CredentialsController credentials.MulticlusterController
-
-	GatewayAPIController GatewayController
-
-	// EndpointShards for a service. This is a global (per-server) list, built from
-	// incremental updates. This is keyed by service and namespace
-	EndpointIndex *EndpointIndex
-
-	// Cache for XDS resources.
-	Cache XdsCache
-}
-
-func (e *Environment) Mesh() *meshconfig.MeshConfig {
-	if e != nil && e.Watcher != nil {
-		return e.Watcher.Mesh()
-	}
-	return nil
-}
-
-func (e *Environment) MeshNetworks() *meshconfig.MeshNetworks {
-	if e != nil && e.NetworksWatcher != nil {
-		return e.NetworksWatcher.Networks()
-	}
-	return nil
-}
-
-// SetPushContext sets the push context with lock protected
-func (e *Environment) SetPushContext(pc *PushContext) {
-	e.mutex.Lock()
-	defer e.mutex.Unlock()
-	e.pushContext = pc
-}
-
-// PushContext returns the push context with lock protected
-func (e *Environment) PushContext() *PushContext {
-	e.mutex.RLock()
-	defer e.mutex.RUnlock()
-	return e.pushContext
-}
-
-// GetDiscoveryAddress parses the DiscoveryAddress specified via MeshConfig.
-func (e *Environment) GetDiscoveryAddress() (host.Name, string, error) {
-	proxyConfig := mesh.DefaultProxyConfig()
-	if e.Mesh().DefaultConfig != nil {
-		proxyConfig = e.Mesh().DefaultConfig
-	}
-	hostname, port, err := net.SplitHostPort(proxyConfig.DiscoveryAddress)
-	if err != nil {
-		return "", "", fmt.Errorf("invalid Istiod Address: %s, %v", proxyConfig.DiscoveryAddress, err)
-	}
-	if _, err := strconv.Atoi(port); err != nil {
-		return "", "", fmt.Errorf("invalid Istiod Port: %s, %s, %v", port, proxyConfig.DiscoveryAddress, err)
-	}
-	return host.Name(hostname), port, nil
-}
-
-func (e *Environment) AddMeshHandler(h func()) {
-	if e != nil && e.Watcher != nil {
-		e.Watcher.AddMeshHandler(h)
-	}
-}
-
-func (e *Environment) AddNetworksHandler(h func()) {
-	if e != nil && e.NetworksWatcher != nil {
-		e.NetworksWatcher.AddNetworksHandler(h)
-	}
-}
-
-func (e *Environment) AddMetric(metric monitoring.Metric, key string, proxyID, msg string) {
-	if e != nil {
-		e.PushContext().AddMetric(metric, key, proxyID, msg)
-	}
-}
-
-// Init initializes the Environment for use.
-func (e *Environment) Init() {
-	// Use a default DomainSuffix, if none was provided.
-	if len(e.DomainSuffix) == 0 {
-		e.DomainSuffix = constants.DefaultClusterLocalDomain
-	}
-
-	// Create the cluster-local service registry.
-	e.clusterLocalServices = NewClusterLocalProvider(e)
-}
-
-func (e *Environment) InitNetworksManager(updater XDSUpdater) (err error) {
-	e.NetworkManager, err = NewNetworkManager(e, updater)
-	return
-}
-
-func (e *Environment) ClusterLocal() ClusterLocalProvider {
-	return e.clusterLocalServices
-}
-
-func (e *Environment) GetProxyConfigOrDefault(ns string, labels, annotations map[string]string, meshConfig *meshconfig.MeshConfig) *meshconfig.ProxyConfig {
-	push := e.PushContext()
-	if push != nil && push.ProxyConfigs != nil {
-		if generatedProxyConfig := push.ProxyConfigs.EffectiveProxyConfig(
-			&NodeMetadata{
-				Namespace:   ns,
-				Labels:      labels,
-				Annotations: annotations,
-			}, meshConfig); generatedProxyConfig != nil {
-			return generatedProxyConfig
-		}
-	}
-	return mesh.DefaultProxyConfig()
-}
 
 // Resources is an alias for array of marshaled resources.
 type Resources = []*discovery.Resource
