+++ pilot/pkg/model/service.go
@@ -936,110 +936,7 @@ type ServiceDiscovery interface {
 	// Kubernetes Multi-Cluster Services (MCS) ServiceExport API. Only applies to services in
 	// Kubernetes clusters.
 	MCSServices() []MCSServiceInfo
-	AmbientIndexes
 }
-
-type AmbientIndexes interface {
-	ServicesWithWaypoint(key string) []ServiceWaypointInfo
-	AddressInformation(addresses sets.String) ([]AddressInfo, sets.String)
-	AdditionalPodSubscriptions(
-		proxy *Proxy,
-		allAddresses sets.String,
-		currentSubs sets.String,
-	) sets.String
-	Policies(requested sets.Set[ConfigKey]) []WorkloadAuthorization
-	ServicesForWaypoint(WaypointKey) []ServiceInfo
-	WorkloadsForWaypoint(WaypointKey) []WorkloadInfo
-}
-
-// WaypointKey is a multi-address extension of NetworkAddress which is commonly used for lookups in AmbientIndex
-// We likely need to consider alternative keying options internally such as hostname as we look to expand beyond istio-waypoint
-// This extension can ideally support that type of lookup in the interface without introducing scope creep into things
-// like NetworkAddress
-type WaypointKey struct {
-	Namespace string
-	Hostnames []string
-
-	Network   string
-	Addresses []string
-}
-
-// WaypointKeyForProxy builds a key from a proxy to lookup
-func WaypointKeyForProxy(node *Proxy) WaypointKey {
-	return waypointKeyForProxy(node, false)
-}
-
-func WaypointKeyForNetworkGatewayProxy(node *Proxy) WaypointKey {
-	return waypointKeyForProxy(node, true)
-}
-
-func waypointKeyForProxy(node *Proxy, externalAddresses bool) WaypointKey {
-	key := WaypointKey{
-		Namespace: node.ConfigNamespace,
-		Network:   node.Metadata.Network.String(),
-	}
-	for _, svct := range node.ServiceTargets {
-		key.Hostnames = append(key.Hostnames, svct.Service.Hostname.String())
-
-		var ips []string
-		if externalAddresses {
-			ips = svct.Service.Attributes.ClusterExternalAddresses.GetAddressesFor(node.GetClusterID())
-		} else {
-			ips = svct.Service.ClusterVIPs.GetAddressesFor(node.GetClusterID())
-		}
-		// if we find autoAllocated addresses then ips should contain constants.UnspecifiedIP which should not be used
-		foundAutoAllocated := false
-		if svct.Service.AutoAllocatedIPv4Address != "" {
-			key.Addresses = append(key.Addresses, svct.Service.AutoAllocatedIPv4Address)
-			foundAutoAllocated = true
-		}
-		if svct.Service.AutoAllocatedIPv6Address != "" {
-			key.Addresses = append(key.Addresses, svct.Service.AutoAllocatedIPv6Address)
-			foundAutoAllocated = true
-		}
-		if !foundAutoAllocated {
-			key.Addresses = append(key.Addresses, ips...)
-		}
-	}
-	return key
-}
-
-// NoopAmbientIndexes provides an implementation of AmbientIndexes that always returns nil, to easily "skip" it.
-type NoopAmbientIndexes struct{}
-
-func (u NoopAmbientIndexes) AddressInformation(sets.String) ([]AddressInfo, sets.String) {
-	return nil, nil
-}
-
-func (u NoopAmbientIndexes) AdditionalPodSubscriptions(
-	*Proxy,
-	sets.String,
-	sets.String,
-) sets.String {
-	return nil
-}
-
-func (u NoopAmbientIndexes) Policies(sets.Set[ConfigKey]) []WorkloadAuthorization {
-	return nil
-}
-
-func (u NoopAmbientIndexes) ServicesForWaypoint(WaypointKey) []ServiceInfo {
-	return nil
-}
-
-func (u NoopAmbientIndexes) Waypoint(string, string) []netip.Addr {
-	return nil
-}
-
-func (u NoopAmbientIndexes) WorkloadsForWaypoint(WaypointKey) []WorkloadInfo {
-	return nil
-}
-
-func (u NoopAmbientIndexes) ServicesWithWaypoint(string) []ServiceWaypointInfo {
-	return nil
-}
-
-var _ AmbientIndexes = NoopAmbientIndexes{}
 
 type AddressInfo struct {
 	*workloadapi.Address
