+++ pilot/pkg/xds/debug.go
@@ -974,7 +974,7 @@ func (s *DiscoveryServer) instancesz(w http.ResponseWriter, req *http.Request) {
 }
 
 func (s *DiscoveryServer) ambientz(w http.ResponseWriter, req *http.Request) {
-	addresses, _ := s.Env.ServiceDiscovery.AddressInformation(nil)
+	addresses, _ := s.Env.AddressInformation(nil)
 	res := struct {
 		Workloads []jsonMarshalProto `json:"workloads"`
 		Services  []jsonMarshalProto `json:"services"`
@@ -1021,7 +1021,7 @@ func (s *DiscoveryServer) ambientz(w http.ResponseWriter, req *http.Request) {
 			res.Services = append(res.Services, jsonMarshalProto{s})
 		}
 	}
-	for _, policy := range s.Env.ServiceDiscovery.Policies(nil) {
+	for _, policy := range s.Env.Policies(nil) {
 		res.Policies = append(res.Policies, jsonMarshalProto{policy.Authorization})
 	}
 	writeJSON(w, res, req)
